@@ -1,8 +1,5 @@
 module github.com/cgrates/birpc
 
-go 1.16
+go 1.23
 
-require (
-	github.com/cenk/hub v1.0.1 // indirect
-	github.com/cenkalti/hub v1.0.1
-)
+require github.com/cenkalti/hub v1.0.2
